+++ frontend/src/components/ui/aspect-ratio.tsx
