+++ frontend/src/components/ui/resizable.tsx
